Purge tables & Remove tables

+++ admin/models/maintImageFiles.php
@@ -16,17 +16,17 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 	protected $imagePath_original;
 	protected $imagePath_watermarked;
 
-    protected function removeAllImageFiles ()
+    public function removeAllImageFiles ()
     {
         $msg = "Remove image files: \n";
 
 		// assign class variables
 		$this->getImagePaths (); 		
 		
-		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_thumb);
-		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_display);
-		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_original);
-		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_watermarked);
+		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_thumb) . '<br>';
+		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_display) . '<br>';
+		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_original) . '<br>';
+		$msg .=  $this->RemoveImagesInFolder ($this->imagePath_watermarked) . '<br>';
 
         return $msg;
     }
@@ -43,16 +43,35 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 			$msg .= "\n" . $errMsg;  
 			
 			JFactory::getApplication()->enqueueMessage($msg, 'error');
+			return $msg;
 		}
-		
-		// ToDo: check that path is valid and not a base path to "anywhere"
-		
-		
+
+		// Check that path is valid
+		if(! is_dir ($fullPath))
+		{
+			$errMsg = JText::_('COM_RSGALLERY2_FOLDER_DOES_NOT_EXIST') . ': "' . $fullPath . '""';
+			$msg .= "\n" . $errMsg;
+
+			JFactory::getApplication()->enqueueMessage($msg, 'error');
+			return $msg;
+		}
+
+		/* ToDo: check that path is valid and not a base path to "anywhere"
+		if(! ($fullPath == JPATH_ROOT))
+		{
+			$errMsg = JText::_('COM_RSGALLERY2_FOLDER_DOES_NOT_EXIST') . ': "' . $fullPath . '""';
+			$msg .= "\n" . $errMsg;
+
+			JFactory::getApplication()->enqueueMessage($msg, 'error');
+			return $msg;
+		}
+		/**/
+
 		//--- remove display images ------------------------
 
 		try
 		{
-			foreach ( glob( $fullPath.'*' ) as $filename ) {
+			foreach ( glob( $fullPath.'\*' ) as $filename ) {
 				if( is_file( $filename )) unlink( $filename );
 			}
 			
@@ -65,10 +84,16 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 		return $msg;
 	}
 
-	// ToDo: Move to model image pathe and inherit from there or rename this class ...
+	// ToDo: Move to model image path and inherit from there or rename this class ...
     private function getImagePaths ()
     {
         // ToDo: Throws .... \Jdatabaseexception ....
+
+	    // preset if following code fails
+	    $this->imagePath_thumb = '';
+	    $this->imagePath_display = '';
+	    $this->imagePath_original = '';
+	    $this->imagePath_watermarked = '';
 
 		//--- thumb -------------------------------------
         $db = JFactory::getDbo();
@@ -79,10 +104,14 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 			->where($db->quoteName('name')." = ".$db->quote('imgPath_thumb'));
 		
 		$db->setQuery($query);
-		$this->imagePath_thumb  = $db->loadResult();
-		
-		//--- display -------------------------------------
-        $db = JFactory::getDbo();
+	    $Path = $db->loadResult();
+	    if(strlen(trim($Path)) > 0)
+	    {
+		    $this->imagePath_thumb = JPATH_ROOT . $Path;
+		}
+
+	    //--- display -------------------------------------
+	    $db = JFactory::getDbo();
 		$query = $db->getQuery (true);
 
 		$query->select ($db->quoteName('value'))
@@ -90,8 +119,12 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 			->where($db->quoteName('name')." = ".$db->quote('imgPath_display'));
 		
 		$db->setQuery($query);
-		$this->imagePath_thumb  = $db->loadResult();
-		
+	    $Path = $db->loadResult();
+		if(strlen(trim($Path)) > 0)
+		{
+			$this->imagePath_display = JPATH_ROOT . $Path;
+		}
+
 		//--- original -------------------------------------
         $db = JFactory::getDbo();
 		$query = $db->getQuery (true);
@@ -101,20 +134,26 @@ class Rsgallery2ModelMaintImageFiles extends  JModelList
 			->where($db->quoteName('name')." = ".$db->quote('imgPath_original'));
 		
 		$db->setQuery($query);
-		$this->imagePath_thumb  = $db->loadResult();
-		
+	    $Path = $db->loadResult();
+		if(strlen(trim($Path)) > 0)
+		{
+			$this->imagePath_original = JPATH_ROOT . $Path;
+		}
+
 		//--- water marked -------------------------------------
         $db = JFactory::getDbo();
 		$query = $db->getQuery (true);
 
-		
 		$query->select ($db->quoteName('value'))
 			->from($db->quoteName('#__rsgallery2_config'))
 			->where($db->quoteName('name')." = ".$db->quote('imgPath_watermarked'));
 		
 		$db->setQuery($query);
-		$this->imagePath_thumb  = $db->loadResult();
-
+	    $Path = $db->loadResult();
+		if(strlen(trim($Path)) > 0)
+		{
+			$this->imagePath_watermarked = JPATH_ROOT . $Path;
+		}
 
 		/* ToDo ?
         if($db->getErrorMsg()){
